+++ src/main/java/org/orekit/utils/MultipleShooter.java
@@ -128,7 +128,7 @@ public class MultipleShooter extends AbstractMultipleShooting {
         int i = 0;
 
         for (int index : mapConstraints.keySet()) {
-            final int np = (int) (index / 6);
+            final int np = index / 6;
             final int nc = index % 6;
             final AbsolutePVCoordinates absPv = patchedSpacecraftStates.get(np).getAbsPVA();
             if (nc < 3) {
